New system

+++ src/chunks_processing.js
@@ -1,30 +1,47 @@
 import csvToJsonFormat from './models/csvFormat.js';
 import fs from 'fs';
-import * as mongoUtil from './mongodb_utils.js';
+import { mongoUrl } from './constants/constants.js';
+import { MongoClient } from 'mongodb';
 
 // Processes the given file.
 const processChunk = (folder, chunkFile) => {
-  const data = fs.readFileSync(`${folder}${chunkFile}`);
-  const chunkJson = parseChunk(data);
-  console.log('data len' + data.length);
-  const mongodb = mongoUtil.getDb();
-  mongodb.collection('sirene').insertMany(chunkJson);
+  if (!fs.existsSync(`${folder}${chunkFile}`)) {
+    return;
+  }
 
-  //Delete the  file when processing is done.
-  fs.unlink(`${folder}${chunkFile}`, (err) => {
-    if (err) console.log(err);
-  });
+  MongoClient.connect(
+    mongoUrl,
+    { useNewUrlParser: true },
+    function (err, client) {
+      if (err) throw err;
+      if (!fs.existsSync(`${folder}${chunkFile}`)) {
+        return;
+      }
+
+      const db = client.db('sirene_invider');
+      const bulk = db.collection('sirene').initializeUnorderedBulkOp();
+      const data = fs.readFileSync(`${folder}${chunkFile}`);
+      parseChunk(bulk, data);
+
+      // Execute the operations
+      bulk.execute();
+
+      //Delete the  file when processing is done.
+      fs.unlink(`${folder}${chunkFile}`, (err) => {
+        if (err) console.log(`Delete: ${err}`);
+      });
+    }
+  );
 };
 
 // Returns a JSON object from the given data.
-const parseChunk = (data) => {
-  var dateJson = [];
+const parseChunk = (bulk, data) => {
   const lines = Buffer.from(data).toString().split('\n');
   lines.shift();
-  for (let index = 0; index < lines.length; index += 1) {
-    dateJson.push(csvToJsonFormat(lines[index].split(',')));
+  let count = 0;
+  for (let index = 0; index < lines.length; index += 1, count += 1) {
+    bulk.insert(csvToJsonFormat(lines[index].split(',')));
   }
-  return dateJson;
 };
 
 export default processChunk;
